fix: load installation before validating admin user 'json_path'
  ... fixing issue similar to #1017.

refactor: hoist helpers from 'cli.room_authz' to 'cli.cli_util'.
